Manipulação da DatabaseException

+++ src/main/java/br/com/leovieira/dscommerce/services/ProductService.java
@@ -1,14 +1,19 @@
 package br.com.leovieira.dscommerce.services;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Propagation;
 import org.springframework.transaction.annotation.Transactional;
 
 import br.com.leovieira.dscommerce.dto.ProductDTO;
 import br.com.leovieira.dscommerce.entities.Product;
 import br.com.leovieira.dscommerce.repositories.ProductRepository;
+import br.com.leovieira.dscommerce.services.exceptions.DatabaseException;
+import br.com.leovieira.dscommerce.services.exceptions.ResourceNotFoundException;
+import jakarta.persistence.EntityNotFoundException;
 
 @Service
 public class ProductService {
@@ -17,14 +22,15 @@ public class ProductService {
 	
 	@Transactional(readOnly = true)
 	public ProductDTO findById(Long id) {
-		Product p = repository.findById(id).get();
+		Product p = repository.findById(id).orElseThrow(
+				()-> new ResourceNotFoundException("Recurso não encontrado")
+		);
 		return new ProductDTO(p);
 	}
 	
 	@Transactional(readOnly = true)
 	public Page<ProductDTO> findAll(Pageable pageable) {
 		Page<Product> result = repository.findAll(pageable);
-		
 		return result.map(x -> new ProductDTO(x));
 	}
 	
@@ -38,15 +44,27 @@ public class ProductService {
 	
 	@Transactional
 	public ProductDTO update(Long id, ProductDTO dto) {
-		Product entity = repository.getReferenceById(id);
-		copyDtoToEntity(dto, entity);
-		entity = repository.save(entity);
-		return new ProductDTO(entity);
+		try {
+			Product entity = repository.getReferenceById(id);
+			copyDtoToEntity(dto, entity);
+			entity = repository.save(entity);
+			return new ProductDTO(entity);
+		}catch (EntityNotFoundException e) {
+			throw new ResourceNotFoundException("Recurso não encontrado");
+		}
 	}
 	
-	@Transactional
+	@Transactional(propagation = Propagation.SUPPORTS)
 	public void delete(Long id) {
-		repository.deleteById(id);
+		if(!repository.existsById(id)) {
+			throw new ResourceNotFoundException("Recurso não encontrado");
+		}
+		try {
+			repository.deleteById(id);
+		}catch(DataIntegrityViolationException e) {
+			throw new DatabaseException("Falha de integridade referencial");
+		}
+		
 	}
 
 	private void copyDtoToEntity(ProductDTO dto, Product entity) {
